add vitest coverage for measurement data
