+++ docs/feature.md
@@ -47,8 +47,15 @@
 - `npm run build` で TypeScript チェックと Vite build を実行する。
 - `npm run export` で GitHub Pages 用の production build を実行する。
 - `npm run lint` で ESLint を実行する。
+- `npm test` で Vitest による自動テストを実行する。
 - `npm run preview` で build 結果をプレビューする。
 - `npm run preview:pages` で build 結果を GitHub Pages と同じ `/myweight/` 配下として localhost で配信する。
+
+## 自動テスト
+
+- Vitest で自動テストを実行する。
+- HealthPlanet のレスポンスレコードをチャート用データへ変換する純粋関数をテストする。
+- 測定種別タグ、空データセット、日付整形、未知タグの無視、API レスポンス順からチャート表示順への反転を確認する。
 
 ## 未実装・現時点で存在しない機能
 
@@ -58,4 +65,3 @@
 - HealthPlanet 認可フローは実装されていない。
 - バックエンド API はない。
 - 永続化用の独自データベースはない。
-- 自動テストは現時点で用意されていない。
