+++ src/components/WeightData.tsx
@@ -1,55 +1,16 @@
 import axios from 'axios';
 import brotliPromise from 'brotli-dec-wasm';
 import moment from 'moment';
-
-export type MeasurementKey = 'weight' | 'bodyFat';
-
-export type MeasurementMetric = {
-  key: MeasurementKey;
-  label: string;            
-  shortLabel: string;
-  tag: string;
-  unit: string;
-  color: string;
-  domain: [number | 'auto', number | 'auto'];
-};
-
-export type MeasurementData = {
-  date: string;
-  value: number;
-};
-
-export type MeasurementDataSet = Record<MeasurementKey, MeasurementData[]>;
+import {
+  createEmptyMeasurementDataSet,
+  createMeasurementDataSetFromRecords,
+  MeasurementData,
+  MeasurementDataSet,
+  measurementMetrics,
+} from '../lib/measurementData';
 
 export let measurementFetchError: string | null = null;
 
-export const measurementMetrics: MeasurementMetric[] = [
-  {
-    key: 'weight',
-    label: '体重',
-    shortLabel: 'Weight',
-    tag: '6021',
-    unit: 'kg',
-    color: '#2563eb',
-    domain: [70, 'auto'],
-  },
-  {
-    key: 'bodyFat',
-    label: '体脂肪率',
-    shortLabel: 'Body Fat',
-    tag: '6022',
-    unit: '%',
-    color: '#0f766e',
-    domain: ['auto', 'auto'],
-  },
-];
-
-const parseDate = (dateString: string): string => {
-  const year = parseInt(dateString.slice(0, 4), 10);
-  const month = parseInt(dateString.slice(4, 6), 10) - 1;
-  const day = parseInt(dateString.slice(6, 8), 10);
-  return moment(new Date(year, month, day)).format('YYYY/MM/DD');
-}
 const formatToApiDate = (date: moment.Moment): string => {
   return date.format('YYYYMMDDHHmmss');
 };
@@ -69,13 +30,6 @@ const corsProxyRequestHeaderOverrides: Record<string, string> = {
   'content-type': FORM_URLENCODED_CONTENT_TYPE,
   'accept-encoding': 'identity',
 };
-
-const metricByTag = new Map(measurementMetrics.map(metric => [metric.tag, metric]));
-
-const createEmptyMeasurementDataSet = (): MeasurementDataSet => ({
-  weight: [],
-  bodyFat: [],
-});
 
 const maskAccessToken = (accessToken: string): string =>
   accessToken.length > 8 ? `***${accessToken.slice(-8)}` : '***';
@@ -408,19 +362,7 @@ const fetchInnerScanDataSet = async (
           return createEmptyMeasurementDataSet();
       }
 
-      const result = createEmptyMeasurementDataSet();
-      for (const record of [...records].reverse()) {
-          const metric = metricByTag.get(record.tag);
-          if (!metric) {
-            continue;
-          }
-
-          result[metric.key].push({
-              date: parseDate(record.date),
-              value: Number(record.keydata),
-          });
-      }
-      return result;
+      return createMeasurementDataSetFromRecords(records);
   } catch (error) {
       console.error('Error fetching measurement data:', error);
       if (axios.isAxiosError(error)) {
